cleaned up test impl
